Fix training loop. Created dir for each example
